seperation model

+++ src/latent_diffusion/modules/diffusionmodules/util.py
@@ -239,7 +239,8 @@ class SiLU(nn.Module):
 
 class GroupNorm32(nn.GroupNorm):
     def forward(self, x):
-        return super().forward(x.float()).type(x.dtype)
+        # return super().forward(x.float()).type(x.dtype)
+        return super().forward(x.float())
 
 
 def conv_nd(dims, *args, **kwargs):
